+++ lib/docassemble/docassemble_api.js
@@ -14,7 +14,7 @@ let da = {};
 module.exports = da;
 
 
-// node -e 'require("./lib/utils/docassemble_api.js").get_dev_id()'
+// node -e 'require("./lib/docassemble/docassemble_api.js").get_dev_id()'
 da.get_dev_id = async function ( timeout ) {
   /* Get docassemble developer id. */
   let options = {
@@ -28,7 +28,7 @@ da.get_dev_id = async function ( timeout ) {
 };  // Ends da.get_dev_id()
 
 
-// node -e 'require("./lib/utils/docassemble_api.js").create_project( "create1" )'
+// node -e 'require("./lib/docassemble/docassemble_api.js").create_project( "create1" )'
 da.create_project = async function ( project_name, timeout ) {
   /* Create a new project on the user's da server. */
   let options = {
@@ -46,7 +46,7 @@ da.create_project = async function ( project_name, timeout ) {
 };  // Ends da.create_project()
 
 
-// node -e 'require("./lib/utils/docassemble_api.js").pull()'
+// node -e 'require("./lib/docassemble/docassemble_api.js").pull()'
 da.pull = async function ( timeout ) {
   /* Pull github code into an existing Project. */
   let options = {
@@ -65,7 +65,7 @@ da.pull = async function ( timeout ) {
 };  // Ends da.pull()
 
 
-// node -e 'require("./lib/utils/docassemble_api.js").has_task_finished()'
+// node -e 'require("./lib/docassemble/docassemble_api.js").has_task_finished()'
 da.has_task_finished = async function ( task_id, timeout ) {
   /* Check task status based on `task_id`. */
   let options = {
@@ -82,7 +82,7 @@ da.has_task_finished = async function ( task_id, timeout ) {
 };  // Ends da.has_task_finished()
 
 
-// node -e 'require("./lib/utils/docassemble_api.js").delete_project( "create1" )'
+// node -e 'require("./lib/docassemble/docassemble_api.js").delete_project( "create1" )'
 da.delete_project = async function ( timeout ) {
   /* Delete a project on the user's da server. */
   let options = {
