+++ lib/docassemble/docassemble_api_interface.js
@@ -1,4 +1,4 @@
-const da = require('../utils/docassemble_api');
+const da = require('./docassemble_api');
 const log = require('../utils/log');
 const session_vars = require('../utils/session_vars' );
 const time = require('../utils/time');
@@ -38,7 +38,7 @@ dai.server_is_busy = async function ( options ) {
   try {
     let response = await da.get_dev_id( timeout );
     if ( response && response.data && response.data.id ) {
-      log.debug({ pre: `Server responded.` });
+      log.info({ pre: `Server responded.` });
       return false;
     }
