--- lib/github_integration/takedown.js
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-
-const install_branch = require('./install-branch.js');
-try {
-  install_branch.takedown();
-} catch ( error ) {
-  throw error;
-}
